move components to correct directory
